Move utility Tensor functions to misc_ops module
* also deletes duplicated func
